+++ xenian_bot/commands/__init__.py
@@ -12,3 +12,4 @@ from .calculator import *
 from .group_management import *
 from .google import *
 from .yandex import *
+from .image_to_text import *